[Feature] Support DDP Evaluation for IS and FID (#151)
* support dist eval for IS and FID

* support offline mode with DDP

* update readme and script for slurm

* fix bug in log print

* fix bug in image saving

* fix bug of log display in offline evaluation

+++ mmgen/core/evaluation/evaluation.py
@@ -6,6 +6,8 @@ from copy import deepcopy
 
 import mmcv
 import torch
+import torch.distributed as dist
+from mmcv.runner import get_dist_info
 from prettytable import PrettyTable
 from torchvision.utils import save_image
 
@@ -33,7 +35,7 @@ def make_metrics_table(train_cfg, ckpt, eval_info, metrics):
     return table.get_string()
 
 
-def make_vanilla_dataloader(img_path, batch_size):
+def make_vanilla_dataloader(img_path, batch_size, dist=False):
     pipeline = [
         dict(type='LoadImageFromFile', key='real_img', io_backend='disk'),
         dict(
@@ -55,24 +57,24 @@ def make_vanilla_dataloader(img_path, batch_size):
         dataset,
         samples_per_gpu=batch_size,
         workers_per_gpu=4,
-        dist=False,
+        dist=dist,
         shuffle=True)
     return dataloader
 
 
 @torch.no_grad()
-def single_gpu_evaluation(model,
-                          data_loader,
-                          metrics,
-                          logger,
-                          basic_table_info,
-                          batch_size,
-                          samples_path=None,
-                          **kwargs):
-    """Evaluate model with a single gpu.
+def offline_evaluation(model,
+                       data_loader,
+                       metrics,
+                       logger,
+                       basic_table_info,
+                       batch_size,
+                       samples_path=None,
+                       **kwargs):
+    """Evaluate model in offline mode.
 
-    This method evaluate model with a single gpu and displays eval progress
-     bar.
+    This method first save generated images at local and then load them by
+    dataloader.
 
     Args:
         model (nn.Module): Model to be tested.
@@ -87,11 +89,13 @@ def single_gpu_evaluation(model,
             evaluation. Default to None.
         kwargs (dict): Other arguments.
     """
-    # eval special and probabilistic metric online only
+    # eval special and recon metric online only
     online_metric_name = ['PPL', 'GaussianKLD']
     for metric in metrics:
         assert metric.name not in online_metric_name, 'Please eval '\
              f'{metric.name} online'
+
+    rank, ws = get_dist_info()
 
     delete_samples_path = False
     if samples_path:
@@ -118,14 +122,15 @@ def single_gpu_evaluation(model,
         max_num_images = max(metric.num_images for metric in metrics)
     num_needed = max(max_num_images - num_exist, 0)
 
-    if num_needed > 0:
+    if num_needed > 0 and rank == 0:
         mmcv.print_log(f'Sample {num_needed} fake images for evaluation',
                        'mmgen')
         # define mmcv progress bar
         pbar = mmcv.ProgressBar(num_needed)
 
-    # if no images, `num_exist` should be zero
-    for begin in range(num_exist, num_needed, batch_size):
+    # if no images, `num_needed` should be zero
+    total_batch_size = batch_size * ws
+    for begin in range(0, num_needed, total_batch_size):
         end = min(begin + batch_size, max_num_images)
         fakes = model(
             None,
@@ -133,7 +138,15 @@ def single_gpu_evaluation(model,
             return_loss=False,
             sample_model=basic_table_info['sample_model'],
             **kwargs)
-        pbar.update(end - begin)
+        global_end = min(begin + total_batch_size, max_num_images)
+        if rank == 0:
+            pbar.update(global_end - begin)
+
+        # gather generated images
+        if ws > 1:
+            placeholder = [torch.zeros_like(fakes) for _ in range(ws)]
+            dist.all_gather(placeholder, fakes)
+            fakes = torch.cat(placeholder, dim=0)
 
         # save as three-channel
         if fakes.size(1) == 3:
@@ -145,14 +158,15 @@ def single_gpu_evaluation(model,
                                'channels in the first dimension, '
                                'not %d' % fakes.size(1))
 
-        for i in range(end - begin):
-            images = fakes[i:i + 1]
-            images = ((images + 1) / 2)
-            images = images.clamp_(0, 1)
-            image_name = str(begin + i) + '.png'
-            save_image(images, os.path.join(samples_path, image_name))
+        if rank == 0:
+            for i in range(global_end - begin):
+                images = fakes[i:i + 1]
+                images = ((images + 1) / 2)
+                images = images.clamp_(0, 1)
+                image_name = str(num_exist + begin + i) + '.png'
+                save_image(images, os.path.join(samples_path, image_name))
 
-    if num_needed > 0:
+    if num_needed > 0 and rank == 0:
         sys.stdout.write('\n')
 
     # return if only save sampled images
@@ -161,13 +175,17 @@ def single_gpu_evaluation(model,
 
     # empty cache to release GPU memory
     torch.cuda.empty_cache()
-    fake_dataloader = make_vanilla_dataloader(samples_path, batch_size)
+    fake_dataloader = make_vanilla_dataloader(
+        samples_path, batch_size, dist=ws > 1)
     for metric in metrics:
         mmcv.print_log(f'Evaluate with {metric.name} metric.', 'mmgen')
         metric.prepare()
-        # prepare for pbar
-        total_need = metric.num_real_need + metric.num_fake_need
-        pbar = mmcv.ProgressBar(total_need)
+        if rank == 0:
+            # prepare for pbar
+            total_need = (
+                metric.num_real_need + metric.num_fake_need -
+                metric.num_real_feeded - metric.num_fake_feeded)
+            pbar = mmcv.ProgressBar(total_need)
         # feed in real images
         for data in data_loader:
             # key for unconditional GAN
@@ -185,40 +203,46 @@ def single_gpu_evaluation(model,
             if reals.shape[1] == 1:
                 reals = torch.cat([reals] * 3, dim=1)
             num_left = metric.feed(reals, 'reals')
-            pbar.update(reals.shape[0])
             if num_left <= 0:
                 break
+            if rank == 0:
+                pbar.update(reals.shape[0] * ws)
         # feed in fake images
         for data in fake_dataloader:
             fakes = data['real_img']
             if fakes.shape[1] == 1:
                 fakes = torch.cat([fakes] * 3, dim=1)
             num_left = metric.feed(fakes, 'fakes')
-            pbar.update(fakes.shape[0])
             if num_left <= 0:
                 break
-        metric.summary()
-        sys.stdout.write('\n')
-    table_str = make_metrics_table(basic_table_info['train_cfg'],
-                                   basic_table_info['ckpt'],
-                                   basic_table_info['sample_model'], metrics)
-    logger.info('\n' + table_str)
-    if delete_samples_path:
-        shutil.rmtree(samples_path)
+            if rank == 0:
+                pbar.update(fakes.shape[0] * ws)
+        if rank == 0:
+            # only call summary at main device
+            metric.summary()
+            sys.stdout.write('\n')
+    if rank == 0:
+        table_str = make_metrics_table(basic_table_info['train_cfg'],
+                                       basic_table_info['ckpt'],
+                                       basic_table_info['sample_model'],
+                                       metrics)
+        logger.info('\n' + table_str)
+        if delete_samples_path:
+            shutil.rmtree(samples_path)
 
 
 @torch.no_grad()
-def single_gpu_online_evaluation(model, data_loader, metrics, logger,
-                                 basic_table_info, batch_size, **kwargs):
-    """Evaluate model with a single gpu in online mode.
+def online_evaluation(model, data_loader, metrics, logger, basic_table_info,
+                      batch_size, **kwargs):
+    """Evaluate model in online mode.
 
-    This method evaluate model with a single gpu and displays eval progress
-    bar. Different form `single_gpu_evaluation`, this function will not save
+    This method evaluate model and displays eval progress bar.
+    Different form `offline_evaluation`, this function will not save
     the images or read images from disks. Namely, there do not exist any IO
     operations in this function. Thus, in general, `online` mode will achieve a
     faster evaluation. However, this mode will take much more memory cost.
-    Therefore This evaluation function is recommended to evaluate your model
-    with a single metric.
+    To be noted that, we only support distributed evaluation for FID and IS
+    currently.
 
     Args:
         model (nn.Module): Model to be tested.
@@ -240,12 +264,20 @@ def single_gpu_online_evaluation(model, data_loader, metrics, logger,
     # receive images but receive a dict with corresponding probabilistic
     # parameter.
 
+    rank, ws = get_dist_info()
+
     special_metrics = []
     recon_metrics = []
     vanilla_metrics = []
     special_metric_name = ['PPL']
     recon_metric_name = ['GaussianKLD']
     for metric in metrics:
+        if ws > 1:
+            assert metric.name in [
+                'FID', 'IS'
+            ], ('We only support FID and IS for distributed evaluation '
+                f'currently, but receive {metric.name}')
+
         if metric.name in special_metric_name:
             special_metrics.append(metric)
         elif metric.name in recon_metric_name:
@@ -259,9 +291,10 @@ def single_gpu_online_evaluation(model, data_loader, metrics, logger,
         metric.prepare()
         max_num_images = max(max_num_images,
                              metric.num_real_need - metric.num_real_feeded)
-    mmcv.print_log(f'Sample {max_num_images} real images for evaluation',
-                   'mmgen')
-    pbar = mmcv.ProgressBar(max_num_images)
+    if rank == 0:
+        mmcv.print_log(f'Sample {max_num_images} real images for evaluation',
+                       'mmgen')
+        pbar = mmcv.ProgressBar(max_num_images)
 
     # avoid `data_loader` is None
     data_loader = [] if data_loader is None else data_loader
@@ -297,20 +330,27 @@ def single_gpu_online_evaluation(model, data_loader, metrics, logger,
 
         if num_feed <= 0:
             break
-        pbar.update(num_feed)
+
+        if rank == 0:
+            pbar.update(num_feed)
+
+    if rank == 0:
+        # finish the pbar stdout
+        sys.stdout.write('\n')
 
     # define mmcv progress bar
     max_num_images = 0 if len(vanilla_metrics) == 0 else max(
         metric.num_fake_need for metric in vanilla_metrics)
-    mmcv.print_log(f'Sample {max_num_images} fake images for evaluation',
-                   'mmgen')
-    pbar = mmcv.ProgressBar(max_num_images)
+    if rank == 0:
+        mmcv.print_log(f'Sample {max_num_images} fake images for evaluation',
+                       'mmgen')
+        pbar = mmcv.ProgressBar(max_num_images)
     # sampling fake images and directly send them to metrics
-    for begin in range(0, max_num_images, batch_size):
-        end = min(begin + batch_size, max_num_images)
+    total_batch_size = batch_size * ws
+    for _ in range(0, max_num_images, total_batch_size):
         fakes = model(
             None,
-            num_batches=end - begin,
+            num_batches=batch_size,
             return_loss=False,
             sample_model=basic_table_info['sample_model'],
             **kwargs)
@@ -321,17 +361,19 @@ def single_gpu_online_evaluation(model, data_loader, metrics, logger,
                                'not % d' % fakes.shape[1])
         if fakes.shape[1] == 1:
             fakes = torch.cat([fakes] * 3, dim=1)
-        pbar.update(end - begin)
-        fakes = fakes[:end - begin]
 
         for metric in vanilla_metrics:
             # feed in fake images
-            _ = metric.feed(fakes, 'fakes')
+            metric.feed(fakes, 'fakes')
 
-    # finish the pbar stdout
-    sys.stdout.write('\n')
+        if rank == 0:
+            pbar.update(total_batch_size)
 
-    # feed special metric
+    if rank == 0:
+        # finish the pbar stdout
+        sys.stdout.write('\n')
+
+    # feed special metric, we do not consider distributed eval here
     for metric in special_metrics:
         metric.prepare()
         fakedata_iterator = iter(
@@ -350,10 +392,12 @@ def single_gpu_online_evaluation(model, data_loader, metrics, logger,
         # finish the pbar stdout
         sys.stdout.write('\n')
 
-    for metric in metrics:
-        metric.summary()
+    if rank == 0:
+        for metric in metrics:
+            metric.summary()
 
-    table_str = make_metrics_table(basic_table_info['train_cfg'],
-                                   basic_table_info['ckpt'],
-                                   basic_table_info['sample_model'], metrics)
-    logger.info('\n' + table_str)
+        table_str = make_metrics_table(basic_table_info['train_cfg'],
+                                       basic_table_info['ckpt'],
+                                       basic_table_info['sample_model'],
+                                       metrics)
+        logger.info('\n' + table_str)
